refactor: move colors visualization file to visualization folder.
